DRTII-1467 Improve naming
