+++ server/src/main/scala/drt/server/feeds/edi/EdiArrival.scala
@@ -1,31 +1,30 @@
 package drt.server.feeds.edi
 
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
-import drt.server.feeds.Arriveable
+import drt.server.feeds.AzinqArrival
 import spray.json.{DefaultJsonProtocol, RootJsonFormat}
-import uk.gov.homeoffice.drt.arrivals._
 import uk.gov.homeoffice.drt.ports.Terminals
 import uk.gov.homeoffice.drt.ports.Terminals.A2
 import uk.gov.homeoffice.drt.time.SDate
 
 
-case class AzinqEdiArrival(AIBT: Option[String],
-                           AirlineIATA: String,
-                           ALDT: Option[String],
-                           CarouselCode: Option[String],
-                           CodeSharePrimaryFlightId: Option[Int],
-                           DepartureArrivalType: String,
-                           EstimatedDateTime: Option[String],
-                           FlightNumber: String,
-                           FlightStatus: String,
-                           GateCode: Option[String],
-                           MaxPax: Option[Int],
-                           OriginDestAirportIATA: String,
-                           ScheduledDateTime: String,
-                           StandCode: Option[String],
-                           TerminalCode: String,
-                           TotalPassengerCount: Option[Int],
-                          ) extends Arriveable {
+case class EdiArrival(AIBT: Option[String],
+                      AirlineIATA: String,
+                      ALDT: Option[String],
+                      CarouselCode: Option[String],
+                      CodeSharePrimaryFlightId: Option[Int],
+                      DepartureArrivalType: String,
+                      EstimatedDateTime: Option[String],
+                      FlightNumber: String,
+                      FlightStatus: String,
+                      GateCode: Option[String],
+                      MaxPax: Option[Int],
+                      OriginDestAirportIATA: String,
+                      ScheduledDateTime: String,
+                      StandCode: Option[String],
+                      TerminalCode: String,
+                      TotalPassengerCount: Option[Int],
+                     ) extends AzinqArrival {
 
   override val maybeEstimated: Option[Long] = EstimatedDateTime.map(SDate(_).millisSinceEpoch)
   override val maybeEstimatedChox: Option[Long] = None
@@ -40,8 +39,8 @@ case class AzinqEdiArrival(AIBT: Option[String],
 }
 
 object AzinqEdiArrivalJsonFormats extends SprayJsonSupport with DefaultJsonProtocol {
-  implicit val azinqEdiArrivalJsonFormat: RootJsonFormat[AzinqEdiArrival] = jsonFormat(
-    AzinqEdiArrival.apply,
+  implicit val azinqEdiArrivalJsonFormat: RootJsonFormat[EdiArrival] = jsonFormat(
+    EdiArrival.apply,
     "AIBT",
     "AirlineIATA",
     "ALDT",
